Revert back the version number to 0.1
I'm not ready to call this library stable just yet. It used
to be an implementation detail of thingengine-core; we're splitting
it out because it makes it easier to test certain portions of Almond
without a full thingengine-core (w/o duplicating too much code,
or too much mocking) but it's not quite API stable.