Make use of package_version class annotation
In the future, we will want to decouple the version of the JS
code (which is written in thingpedia-version in package.json)
to the version of the device class, so that we can update
metadata and thingtalk code without storing a new version of the
JS code (which is expensive).

The first step in that direction is to support a #[package_version]
annotation of the class. If the annotation is not present, it
defaults to the #[version] annotation instead.