Make BuiltinModule derive from BaseJavascriptModule
Split the functionality related to downloading zip files from Thingpedia
and caching into BaseOnDiskJavascriptModule, of which ThingpediaModuleV2 is
the only subclass.

Having BuiltinModule derive from BaseJavascriptModule gives us
a couple of advantages:
- it brings us closer to a clean class hierarchy of modules, and a
  consistent API
- it provides uniform wrapping behavior around get_*, subscribe_*, etc.
- it introduces subdevices support for builtins

The latter in particular is important because we need to test
subdevices in thingengine-core, and relying on true thingpedia devices
for tests is dangerous and prone to breakages.