Stop converting ThingTalk classes to the old manifest format
The old manifest format is obsolete and cannot handle new annotation
types or new configuration types.
We should move away from it, and make use of classdefs internally.

Unfortunately, some of the code in thingpedia-common-devices,
and quite a bit of the code in thingpedia-api relies on accessing
parts of the manifest through 'this.constructor.metadata', so
we need to copy over some code from ThingTalk to convert those
parts.